dom/tests/mochitest/mouselock/file_screenXYclientXYConst.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 633602</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
  </script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602
  </a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>

  <div id="div"></div>

  <pre id="test">
    <script type="application/javascript">
      /** Test for Bug 633602 **/
      /**
      The mouse is moved when it's not locked
      Then the mouse is moved when it is locked
      When the mouse isn't locked, the clientXY and screenXY should
      have values representing their respectives positions on the screen
      However, when the mouse is locked, they all should be equal to zero
      **/

      SimpleTest.waitForExplicitFinish();

      var pointer = navigator.mozPointer;
      var div;
      var countMoves = 0;
      // The width and height of a div element
      // These values will be used to move the mouse
      var divWidth = 0;
      var divHeight = 0;

      // Hold the values for screenX/Y clientX/Y when the mouse itsn't locked
      var screenXUnlocked = 0;
      var screenYUnlocked = 0;
      var clientXUnlocked = 0;
      var clientYUnlocked = 0;

      var moveUnlocked = function (e) {
        // Mouse should be unlocked here
        is(pointer.isLocked(), false, "Mouse should be unlocked");
        // clientX/clientY should be equal to divWidth/divHeight
        // since the mouse isn't locked
        is(e.clientX, divWidth, "clientX should be dynamic since mouse isn't "
          + " locked");
        is(e.clientY, divHeight, "clientY should be dynamic since mouse isn't "
          + " locked");
        screenXUnlocked = e.screenX;
        screenYUnlocked = e.screenY;
        clientXUnlocked = e.clientX;
        clientYUnlocked = e.clientY;

        // Remove the current mousemove listener from "div" and add a new one
        div.removeEventListener("mousemove", moveUnlocked, false);
        div.addEventListener("mousemove", moveUnlockedAgain, false);

        // Move the mouse to coords divWidth and divHeight
        synthesizeMouse(div, divWidth+10, divHeight+10, {
          type: "mousemove"
        }, window);
      };

      var moveUnlockedAgain = function (e) {
        // Mouse should be unlocked here
        is(pointer.isLocked(), false, "Mouse should be unlocked");
        // clientX/clientY should be equal to divWidth+10/divHeight+10
        // since the mouse isn't locked
        is(e.clientX, divWidth+10, "clientX should be dynamic since "
          + " mouse isn't locked");
        is(e.clientY, divHeight+10, "clientY should be dynamic since "
          + " mouse isn't locked");

        screenXUnlocked = e.screenX;
        screenYUnlocked = e.screenY;
        clientXUnlocked = e.clientX;
        clientYUnlocked = e.clientY;

        // Remove the current mousemove listener from "div" and add a new one
        div.removeEventListener("mousemove", moveUnlockedAgain, false);
        div.addEventListener("mousemove", moveLocked, false);

        pointer.lock(div, function () {
          // Move the mouse to coords divWidth and divHeight
          synthesizeMouse(div, divWidth+20, divHeight+20, {
            type: "mousemove"
          }, window);
        });
      };

      var moveLocked = function (e) {
        // Mouse should be locked here
        ok(pointer.isLocked(), "Mouse should be locked");
        // When mouse lock is enabled clientX, clientY, screenX, and screenY
        // must hold constant values 
        // as if the mouse did not move at all once mouse lock was entered
        is(e.screenX, screenXUnlocked, "screenX must hold constant value "
          + " as if the mouse did not move at all once mouse was lock");
        is(e.screenY, screenYUnlocked, "screenY must hold constant value " 
          +" as if the mouse did not move at all once mouse was lock");
        is(e.clientX, clientXUnlocked, "clientX must hold constant value "
          + " as if the mouse did not move at all once mouse was lock");
        is(e.clientY, clientYUnlocked, "clientY must hold constant value "
          + " as if the mouse did not move at all once mouse was lock");

        div.removeEventListener("mousemove", moveLocked, false);

        document.mozCancelFullScreen();
      };

      document.addEventListener("mozfullscreenchange", function() {
        if (document.mozFullScreenElement === div) {
          // Set the values for divWidth and divHeight
          // It's dividing by two to make sure the value is a valid coord
          // The value is being rounded since it's not possible to move
          // the mouse to broken values
          divWidth = Math.round(div.offsetWidth / 2);
          divHeight = Math.round(div.offsetHeight / 2);
          // Moving the mouse when the pointer is not locked
          synthesizeMouse(div, divWidth, divHeight, {
            type: "mousemove"
          }, window);
        } else {
          SimpleTest.finish();
        }
      }, false);

      function start() {
        div = document.getElementById("div");
        div.addEventListener("mousemove", moveUnlocked, false);

        SimpleTest.waitForFocus(function() {
          div.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
